+++ src/api/tiles.ts
@@ -1,0 +1,208 @@
+import { Api, Context } from '.'
+import { Headers } from '../lib/mbtiles'
+import { TileJSON } from '../lib/tilejson'
+import {
+  getInterpolatedUpstreamTileUrl,
+  getTileHeaders,
+  tileToQuadKey,
+} from '../lib/tiles'
+import { hash } from '../lib/utils'
+import { NotFoundError } from './errors'
+
+function noop() {}
+
+export interface TilesApi {
+  getTile(opts: {
+    tilesetId: string
+    zoom: number
+    x: number
+    y: number
+  }): Promise<{ data: Buffer; headers: Headers }>
+  putTile(opts: {
+    tilesetId: string
+    zoom: number
+    x: number
+    y: number
+    data: Buffer
+    etag?: string
+  }): void
+}
+
+function createTilesApi({
+  api,
+  context,
+}: {
+  api: Pick<Api, 'getTilesetInfo'>
+  context: Context
+}): TilesApi {
+  const { db, upstreamRequestsManager } = context
+
+  function getUpstreamTileUrl({
+    tilesetId,
+    zoom,
+    x,
+    y,
+  }: {
+    tilesetId: string
+    zoom: number
+    x: number
+    y: number
+  }) {
+    const { tilejson, upstreamTileUrls } = api.getTilesetInfo(tilesetId)
+
+    if (!upstreamTileUrls) return
+
+    const upstreamTileUrl = getInterpolatedUpstreamTileUrl({
+      tiles: upstreamTileUrls,
+      scheme: tilejson.scheme,
+      zoom,
+      x,
+      y,
+    })
+
+    return upstreamTileUrl
+  }
+
+  const tilesApi: TilesApi = {
+    async getTile({ tilesetId, zoom, x, y }) {
+      const quadKey = tileToQuadKey({ x, y, zoom })
+
+      const row:
+        | {
+            data: Buffer
+            etag?: string
+            tilejson: string
+          }
+        | undefined = db
+        .prepare<{
+          tilesetId: string
+          quadKey: string
+        }>(
+          'SELECT TileData.data as data, Tile.etag as etag FROM TileData ' +
+            'JOIN Tile ON TileData.tileHash = Tile.tileHash ' +
+            'JOIN Tileset ON Tile.tilesetId = Tileset.id ' +
+            'WHERE Tileset.id = :tilesetId AND Tile.quadKey = :quadKey'
+        )
+        .get({ tilesetId, quadKey })
+
+      async function fetchOnlineResource(): Promise<
+        | {
+            data: Buffer
+            etag?: string
+          }
+        | undefined
+      > {
+        const upstreamTileUrl = getUpstreamTileUrl({
+          tilesetId,
+          zoom,
+          x,
+          y,
+        })
+
+        // TODO: Need to check if we can make online requests too
+        if (upstreamTileUrl) {
+          const response = await upstreamRequestsManager.getUpstream({
+            url: upstreamTileUrl,
+            etag: row?.etag,
+            responseType: 'buffer',
+          })
+
+          if (response) {
+            try {
+              tilesApi.putTile({
+                tilesetId,
+                zoom,
+                x,
+                y,
+                data: response.data,
+                etag: response.etag,
+              })
+            } catch (_err) {
+              // TODO: Handle error here?
+              noop()
+            }
+
+            return { data: response.data, etag: response.etag }
+          }
+        }
+      }
+
+      let tile: { data: Buffer; etag?: string } | undefined
+
+      if (row) {
+        tile = { data: row.data, etag: row.etag }
+        fetchOnlineResource().catch(noop)
+      } else {
+        tile = await fetchOnlineResource()
+      }
+
+      if (!tile) {
+        // TODO: Improve error handling here?
+        throw new NotFoundError(
+          `Tileset id = ${tilesetId}, [${zoom}, ${x}, ${y}]`
+        )
+      }
+
+      return {
+        data: tile.data,
+        // TODO: This never returns a Last-Modified header but seems like the endpoint would want it to if possible?
+        // Would require changing the return type of UpstreamRequestsManager.getUpstream
+        headers: { ...getTileHeaders(tile.data), Etag: tile.etag },
+      }
+    },
+    putTile({ tilesetId, zoom, x, y, data, etag }) {
+      const upstreamTileUrl = getUpstreamTileUrl({
+        tilesetId,
+        zoom,
+        x,
+        y,
+      })
+
+      const quadKey = tileToQuadKey({ x, y, zoom })
+
+      const transaction = db.transaction(() => {
+        const tileHash = hash(data).toString('hex')
+
+        db.prepare<{
+          tileHash: string
+          tilesetId: string
+          data: Buffer
+        }>(
+          'INSERT INTO TileData (tileHash, tilesetId, data) VALUES (:tileHash, :tilesetId, :data)'
+        ).run({ tileHash, tilesetId, data })
+
+        // TODO: Is this still necessary?
+        db.prepare<{
+          etag?: string
+          tilesetId: string
+          upstreamUrl?: string
+        }>(
+          'UPDATE Tileset SET (upstreamUrl) = (:upstreamUrl) WHERE id = :tilesetId'
+        ).run({
+          tilesetId,
+          upstreamUrl: upstreamTileUrl,
+        })
+
+        db.prepare<{
+          etag?: string
+          quadKey: string
+          tileHash: string
+          tilesetId: string
+        }>(
+          'INSERT INTO Tile (etag, quadKey, tileHash, tilesetId) VALUES (:etag, :quadKey, :tileHash, :tilesetId)'
+        ).run({
+          etag,
+          quadKey,
+          tileHash,
+          tilesetId,
+        })
+      })
+
+      transaction()
+    },
+  }
+
+  return tilesApi
+}
+
+export default createTilesApi
